requirements.txt
streamlit>=1.28.0
openai>=1.3.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
numpy>=1.24.0